+++ packages/extension/src/parts/Completion/Completion.js
@@ -1,0 +1,20 @@
+import * as Assert from '../Assert/Assert.js'
+import * as Position from '../Position/Position.js'
+import * as Rpc from '../Rpc/Rpc.js'
+import * as TextDocumentSync from '../TextDocumentSync/TextDocumentSync.js'
+
+/**
+ * @type {vscode.CompletionProvider['provideCompletions']}
+ */
+export const getCompletion = async (textDocument, offset) => {
+  const uri = textDocument.uri
+  Assert.string(uri)
+  await TextDocumentSync.openTextDocuments([textDocument])
+  const tsPosition = Position.getTsPosition(textDocument, offset)
+  const tsResult = await Rpc.invoke('Completion.getCompletion', {
+    file: textDocument.uri,
+    line: tsPosition.line,
+    offset: tsPosition.offset,
+  })
+  return tsResult
+}
