+++ packages/extension/src/parts/Definition/Definition.js
@@ -1,0 +1,14 @@
+import * as Position from '../Position/Position.js'
+import * as Rpc from '../Rpc/Rpc.js'
+import * as TextDocumentSync from '../TextDocumentSync/TextDocumentSync.js'
+
+export const getDefinition = async (textDocument, offset) => {
+  await TextDocumentSync.openTextDocuments([textDocument])
+  const tsPosition = Position.getTsPosition(textDocument, offset)
+  const tsResult = await Rpc.invoke('Definition.getDefinition', {
+    file: textDocument.uri,
+    line: tsPosition.line,
+    offset: tsPosition.offset,
+  })
+  return tsResult
+}
