+++ packages/extension/src/parts/ExtensionHost/ExtensionHostCompletionProviderTypeScript.js
@@ -1,8 +1,5 @@
-import * as Assert from '../Assert/Assert.js'
+import * as Completion from '../Completion/Completion.js'
 import * as GetCompletionFromTsResult from '../GetCompletionFromTsResult/GetCompletionFromTsResult.js'
-import * as Rpc from '../Rpc/Rpc.js'
-import * as TextDocumentSync from '../TextDocumentSync/TextDocumentSync.js'
-import * as Position from '../Position/Position.js'
 
 export const languageId = 'typescript'
 
@@ -10,15 +7,7 @@ export const languageId = 'typescript'
  * @type {vscode.CompletionProvider['provideCompletions']}
  */
 export const provideCompletions = async (textDocument, offset) => {
-  const uri = textDocument.uri
-  Assert.string(uri)
-  await TextDocumentSync.openTextDocuments([textDocument])
-  const tsPosition = Position.getTsPosition(textDocument, offset)
-  const tsResult = await Rpc.invoke('Completion.getCompletion', {
-    file: textDocument.uri,
-    line: tsPosition.line,
-    offset: tsPosition.offset,
-  })
+  const tsResult = await Completion.getCompletion(textDocument, offset)
   const items = GetCompletionFromTsResult.getCompletionFromTsResult(tsResult)
   return items
 }
