+++ packages/node/src/parts/TypeDefinition/TypeDefinition.js
@@ -1,0 +1,6 @@
+import * as TsServerRequests from '../TsServerRequests/TsServerRequests.js'
+
+export const getTypeDefinition = async (params) => {
+  const tsResult = await TsServerRequests.typeDefinition(params)
+  return tsResult
+}
